pyproject.toml
[tool.poetry]
name = "numpy2"
version = "0.1.0"
description = ""
authors = ["BoshraJaber <boshrajaber94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.4.3"
jupyterlab = "^3.2.0"
numpy = "1.21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
